requirements.txt
beautifulsoup4==4.11.2
discord.py==2.2.2
discord_ui==5.1.6
googletrans==3.0.0
lxml==4.9.2
Pillow==9.4.0
python-dotenv==1.0.0
requests==2.28.2
riotwatcher==3.2.4
Unidecode==1.3.6
xmljson==0.2.1
yt_dlp==2023.3.4
